Initial commit
The syntax directive '# syntax=docker/dockerfile-upstream:master' is
used in the Dockerfiles to use variable replacement functionality (e.g.
${version_yq#v}). The os ARG is defined before the FROM instruction, so
said instruction can use the ARG. Defining the ARG after the FROM
instruction would cause the ARG to be unusable by it. The trailing slash
for the last argument to the COPY instruction is required to instruct
Docker to copy files into the debian directory and not create a file
named debian from the source file arguments. The Go compiler is
symlinked from /usr/lib/go-${version_go}/bin/go because that is where
the golang-1.23-go package installs the compiler.

The --no-cache option is appended to the docker_build_options Makefile
variable when I toggle the DEBUG_PLAIN_STDOUT Makefile configuration
because what I'm trying to debug usually can be hidden by Docker's
layer caching. The deb-sign Makefile target expects the GPG key's
passphrase to be entered in interactively and not injected via
environment variable because, when trying to implement the latter, I
found the solution non-trivial.